+++ assets/scss/team-theme.scss
@@ -16,3 +16,13 @@ $secondary-graident: linear-gradient(to right, #FFB800, #FF8A00);
 .bg-secondary-gradieht {
   background: $secondary-graident;
 }
+
+.nav__categories {
+  .category-culture a::after{ 
+    content: '🦄';
+  }
+
+  .category-cafebreak a::after{ 
+    content: '☕️';
+  }
+}
