+++ components/CategoryArticles.vue
@@ -1,7 +1,5 @@
 <script lang="ts" setup>
 import { useI18n } from 'vue-i18n'
-import upperFirst from 'lodash.upperFirst'
-import camelCase from 'lodash.camelCase'
 
 /**
  * Define props
@@ -12,21 +10,28 @@ interface Props {
 const props = defineProps<Props>()
 
 // ----------------------------------------------------------------------------
+// Get articles
 const config = useRuntimeConfig()
 const { data } = await useAsyncData('article', async (nuxtApp) => {
   const { $contentfulClient } = nuxtApp
-  return $contentfulClient.getEntries({
+  const entries = await $contentfulClient.getEntries({
     content_type: config.private.CONTENTFUL_CONTENT_KEY,
     'metadata.tags.sys.id[all]': props.categoryId,
     order: '-sys.createdAt',
     limit: 10,
   })
+  const tags = await $contentfulClient.getTags()
+  return { entries: entries, tags: tags }
 })
-const items = data.value.items
+const entries = data.value.entries.items
 
-const categoryTitle = upperFirst(camelCase(props.categoryId))
-const entryCount = items.length
+// Get category title
+const { $tagParser } = useNuxtApp()
+const parsedTags = $tagParser(data.value.tags.items)
+const categoryTitle = parsedTags.find((item) => item.id === props.categoryId).title
+const entryCount = entries.length
 
+// Setup meta tags
 const { t } = useI18n()
 const categoryHeadline = t('category_headline', { category: categoryTitle })
 const uri = useRoute().path
@@ -55,7 +60,7 @@ useHead({
     <p v-if="entryCount === 0">{{ $t('category_article_empty_state') }}</p>
     <p v-else>{{ $t('category_article_count', { count: entryCount }) }}</p>
 
-    <article v-for="entry in items" :key="entry.sys.id">
+    <article v-for="entry in entries" :key="entry.sys.id">
       <div class="grid">
         <div class="col-6">
           <a class="img" :href="`/article/${entry.fields.slug}`">
